+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.9.1] - 2023-04-26
 
+### Changed
+
+- Improve performance when checking for installed Django apps in distribution package
+
 ### Fixed
 
 - Various minor fixes